hopefully fixed issue with rational numbers

+++ src/papilo/core/Problem.hpp
@@ -398,8 +398,8 @@ class Problem
          return outputPair;
       }
       auto rowvec = matrix.getRowCoefficients( row );
-      REAL minvalue = std::numeric_limits<double>::infinity();
-      REAL maxvalue = -std::numeric_limits<double>::infinity();
+      REAL minvalue = std::numeric_limits<REAL>::infinity();
+      REAL maxvalue = -std::numeric_limits<REAL>::infinity();
       for( int j = 0; j < rowvec.getLength(); ++j )
       {
          int col = rowvec.getIndices()[j];
